リファクタ: PackageToLodDict を別ファイルに抽出
