+++ Runtime/CityImport/AreaSelector/AreaSelectorDataPass.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using PLATEAU.CityImport.Setting;
 using PLATEAU.Dataset;
 using PLATEAU.Native;
 using UnityEngine;
