+++ Runtime/CityImport/AreaSelector/SceneObjs/AreaLodView.cs
@@ -3,6 +3,7 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using PLATEAU.CityImport.Setting;
 using PLATEAU.Dataset;
 using PLATEAU.Util;
 using UnityEditor;
